requirements.txt
Flask
flask-cors
flask-pymongo
bcrypt
numpy
tensorflow
python-dotenv
flask-apscheduler
requests
pymongo[srv]
gunicorn